+++ sdk/communication/communication-administration/src/communicationIdentity/models.ts
@@ -3,13 +3,13 @@
 
 import { PipelineOptions } from "@azure/core-http";
 import { CommunicationUserIdentifier } from "@azure/communication-common";
-import { CommunicationIdentityToken } from "./generated/src/models";
+import { CommunicationIdentityAccessToken } from "./generated/src/models";
 import { WithResponse } from "../common/models";
 
 /**
  * Represents the scope of the token.
  */
-export type TokenScope = "chat" | "voip" | "pstn";
+export type TokenScope = "chat" | "voip";
 
 /**
  * Client options used to configure the CommunicationIdentity API requests.
@@ -19,8 +19,7 @@ export interface CommunicationIdentityOptions extends PipelineOptions {}
 /**
  * The issued token and the user it was issued for.
  */
-export interface CommunicationUserToken
-  extends Pick<CommunicationIdentityToken, "token" | "expiresOn"> {
+export interface CommunicationUserToken extends CommunicationIdentityAccessToken {
   /**
    * Represents the user the token was issued for
    */
@@ -36,3 +35,10 @@ export type CreateUserResponse = WithResponse<CommunicationUserIdentifier>;
  * Represents the response from issuing a token
  */
 export type IssueTokenResponse = WithResponse<CommunicationUserToken>;
+
+/**
+ * Represents the response from creating a user with a token
+ */
+export type CreateUserWithTokenResponse = WithResponse<CommunicationUserToken>;
+
+export { CommunicationIdentityAccessToken } from "./generated/src/models";